+++ src/Products/urban/UrbanEvent.py
@@ -14,6 +14,7 @@ __author__ = """Gauthier BASTIEN <gbastien@commune.sambreville.be>, Stephan GEUL
 __docformat__ = 'plaintext'
 
 from AccessControl import ClassSecurityInfo
+from collective.archetypes.select2.select2widget import MultiSelect2Widget
 from Products.Archetypes.atapi import *
 from zope.interface import implements
 import interfaces
@@ -357,7 +358,7 @@ schema = Schema((
     ),
     LinesField(
         name='mainSignatures',
-        widget=MultiSelectionWidget(
+        widget=MultiSelect2Widget(
             format='checkbox',
             label=_('urban_label_mainSignatures',
                     default='Mainsignatures'),