+++ src/Products/urban/content/licence/Declaration.py
@@ -14,6 +14,7 @@ __author__ = """Gauthier BASTIEN <gbastien@commune.sambreville.be>, Stephan GEUL
 __docformat__ = 'plaintext'
 
 from AccessControl import ClassSecurityInfo
+from collective.archetypes.select2.select2widget import MultiSelect2Widget
 from Products.Archetypes.atapi import *
 from zope.interface import implements
 from Products.urban import interfaces
@@ -36,7 +37,7 @@ schema = Schema((
 
     LinesField(
         name='article',
-        widget=MultiSelectionWidget(
+        widget=MultiSelect2Widget(
             label=_('urban_label_article', default='Article'),
         ),
         schemata='urban_description',